restructure: move the auth form to the features file
